Refactor confirmation mechanism
